backend_people_give_this_logic/requirements_backend.txt
torch
transformers
pillow
olmocr
